requirements.txt
matplotlib~=3.9.2
tqdm~=4.66.5
pandas~=2.2.3
optuna~=4.1.0
numpy~=1.24.1
pillow~=11.0.0
torchvision~=0.15.2+cu118
transformers~=4.45.2
seaborn~=0.13.2
scikit-learn~=1.5.2
imblearn~=0.0
torch~=2.0.1+cu118